+++ packages/eslint-plugin/rules/styleConditions.ts
@@ -1,13 +1,14 @@
 import { AST_NODE_TYPES, TSESTree } from "@typescript-eslint/utils";
 import { createRule } from "../utils.js";
+import { importsNextYak, isStyledOrCssTag } from "./utils.js";
 
 type ImportedNames = {
   styled?: string;
   css?: string;
 };
 
-export const yakStyleConditions = createRule({
-  name: "yak-style-conditions",
+export const styleConditions = createRule({
+  name: "style-conditions",
   meta: {
     type: "suggestion",
     docs: {
@@ -26,27 +27,17 @@ export const yakStyleConditions = createRule({
   },
   defaultOptions: [],
   create: (context) => {
-    /** track the importad names for css and styled from next-yak */
-    const importedNames: ImportedNames = {};
+    const { importedNames, ImportDeclaration } = importsNextYak();
     return {
-      ImportDeclaration(node: TSESTree.ImportDeclaration) {
-        if (node.source.value === "next-yak") {
-          node.specifiers.forEach((specifier) => {
-            if (
-              specifier.type === AST_NODE_TYPES.ImportSpecifier &&
-              specifier.imported.type === AST_NODE_TYPES.Identifier
-            ) {
-              if (specifier.imported.name === "styled") {
-                importedNames.styled = specifier.local.name;
-              } else if (specifier.imported.name === "css") {
-                importedNames.css = specifier.local.name;
-              }
-            }
-          });
-        }
-      },
-      /** All template literals */
+      ImportDeclaration,
       TaggedTemplateExpression(node: TSESTree.TaggedTemplateExpression) {
+        if (
+          importedNames.styled === undefined &&
+          importedNames.css === undefined
+        ) {
+          return;
+        }
+
         if (!node.parent || isStyledOrCssTag(node, importedNames) !== "css") {
           return;
         }
@@ -152,38 +143,6 @@ function findClosestStyledOrCssTag(
   }
 
   return { tag: undefined, needle, type: undefined, params };
-}
-
-function isStyledOrCssTag(
-  node: TSESTree.Node,
-  importedNames: ImportedNames,
-): false | "styled" | "css" {
-  if (node.type !== AST_NODE_TYPES.TaggedTemplateExpression) {
-    return false;
-  }
-  const { tag } = node;
-  // Check for simple styled`` or css``
-  if (
-    tag.type === AST_NODE_TYPES.Identifier &&
-    (tag.name === importedNames.styled || tag.name === importedNames.css)
-  ) {
-    return "css" as const;
-  }
-  // Check for styled.button`` or styled(Component)``
-  if (tag.type === AST_NODE_TYPES.MemberExpression) {
-    return tag.object.type === AST_NODE_TYPES.Identifier &&
-      tag.object.name === importedNames.styled
-      ? ("styled" as const)
-      : false;
-  }
-  // Check for styled(button)``
-  if (tag.type === AST_NODE_TYPES.CallExpression) {
-    return tag.callee.type === AST_NODE_TYPES.Identifier &&
-      tag.callee.name === importedNames.styled
-      ? ("styled" as const)
-      : false;
-  }
-  return false;
 }
 
 /**
